requirements.txt
streamlit
google-genai
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client